Likes (#4)
* refactor(api): like delete

+++ src/constants/queries.js
@@ -1,4 +1,4 @@
-module.exports.DELETE_LIKE = 'DELETE FROM likes WHERE id = $1';
+module.exports.DELETE_LIKE = 'DELETE FROM likes WHERE id = $1 and userId = $2';
 module.exports.GET_HISTORY = 'SELECT * FROM history WHERE userId = $1';
 module.exports.GET_LIKES = 'SELECT * FROM likes WHERE userId = $1';
 module.exports.INSERT_HISTORY =
